# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Ambition Classes coaching institute website
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type NoteClass = typeof NoteClass[keyof typeof NoteClass];


export const NoteClass = {
  NUMBER_8: '8',
  NUMBER_9: '9',
  NUMBER_10: '10',
} as const;

export type NoteSubject = typeof NoteSubject[keyof typeof NoteSubject];


export const NoteSubject = {
  Mathematics: 'Mathematics',
  Science: 'Science',
  English: 'English',
  Hindi: 'Hindi',
  SST: 'SST',
  AI: 'AI',
  Sanskrit: 'Sanskrit',
} as const;

export interface Note {
  id: number;
  title: string;
  class: NoteClass;
  subject: NoteSubject;
  /**
     * Storage object path (e.g. `/objects/uploads/uuid`), or null if no PDF has been uploaded yet ("Coming Soon").
     * @nullable
     */
  pdfObjectPath: string | null;
  uploadDate: string;
}

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  class: NoteClass;
  subject: NoteSubject;
  /** @nullable */
  pdfObjectPath?: string | null;
}

export interface NoteUpdate {
  /** @minLength 1 */
  title?: string;
  class?: NoteClass;
  subject?: NoteSubject;
  /** @nullable */
  pdfObjectPath?: string | null;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  password: string;
}

export interface AdminStatus {
  authenticated: boolean;
}

export interface UploadUrlRequest {
  /**
     * Original file name.
     * @minLength 1
     */
  name: string;
  /**
     * File size in bytes.
     * @minimum 1
     */
  size: number;
  /**
     * MIME type of the file (e.g. `application/pdf`).
     * @minLength 1
     */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export type ListNotesParams = {
class?: NoteClass;
subject?: NoteSubject;
};