# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Ambition Classes coaching institute website
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Notes
    description: Study notes catalog operations
  - name: Admin
    description: Admin authentication operations
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /notes:
    get:
      operationId: listNotes
      tags: [Notes]
      summary: List study notes
      description: Publicly list study notes, optionally filtered by class and/or subject. No authentication required.
      parameters:
        - name: class
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/NoteClass"
        - name: subject
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/NoteSubject"
      responses:
        "200":
          description: List of study notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [Notes]
      summary: Create a study note (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /notes/{id}:
    patch:
      operationId: updateNote
      tags: [Notes]
      summary: Update a study note (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Note not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteNote
      tags: [Notes]
      summary: Delete a study note (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Note not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [Admin]
      summary: Log in as admin using the shared admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStatus"
        "401":
          description: Incorrect password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [Admin]
      summary: Log out of the admin session
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStatus"

  /admin/session:
    get:
      operationId: getAdminSession
      tags: [Admin]
      summary: Check whether the current session is authenticated as admin
      responses:
        "200":
          description: Session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStatus"

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for PDF upload (admin only)
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an uploaded PDF (public, no authentication)
      description: |
        Serves object entities uploaded via presigned URLs. Notes PDFs are
        publicly downloadable without any login.
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir (e.g. `uploads/some-uuid`).
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    NoteClass:
      type: string
      enum: ["8", "9", "10"]

    NoteSubject:
      type: string
      enum:
        - Mathematics
        - Science
        - English
        - Hindi
        - SST
        - AI
        - Sanskrit

    Note:
      type: object
      required: [id, title, class, subject, pdfObjectPath, uploadDate]
      properties:
        id:
          type: integer
        title:
          type: string
        class:
          $ref: "#/components/schemas/NoteClass"
        subject:
          $ref: "#/components/schemas/NoteSubject"
        pdfObjectPath:
          type: ["string", "null"]
          description: Storage object path (e.g. `/objects/uploads/uuid`), or null if no PDF has been uploaded yet ("Coming Soon").
        uploadDate:
          type: string
          format: date-time

    NoteInput:
      type: object
      required: [title, class, subject]
      properties:
        title:
          type: string
          minLength: 1
        class:
          $ref: "#/components/schemas/NoteClass"
        subject:
          $ref: "#/components/schemas/NoteSubject"
        pdfObjectPath:
          type: ["string", "null"]

    NoteUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        class:
          $ref: "#/components/schemas/NoteClass"
        subject:
          $ref: "#/components/schemas/NoteSubject"
        pdfObjectPath:
          type: ["string", "null"]

    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1

    AdminStatus:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `application/pdf`).

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string