# lib/api-zod/src/generated/types/adminStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Ambition Classes coaching institute website
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStatus {
  authenticated: boolean;
}