# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Ambition Classes coaching institute website
 * OpenAPI spec version: 0.1.0
 */
import type { NoteClass } from './noteClass';
import type { NoteSubject } from './noteSubject';

export interface Note {
  id: number;
  title: string;
  class: NoteClass;
  subject: NoteSubject;
  /**
     * Storage object path (e.g. `/objects/uploads/uuid`), or null if no PDF has been uploaded yet ("Coming Soon").
     * @nullable
     */
  pdfObjectPath: string | null;
  uploadDate: Date;
}