# lib/api-zod/src/generated/types/noteClass.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Ambition Classes coaching institute website
 * OpenAPI spec version: 0.1.0
 */

export type NoteClass = typeof NoteClass[keyof typeof NoteClass];


export const NoteClass = {
  NUMBER_8: '8',
  NUMBER_9: '9',
  NUMBER_10: '10',
} as const;