# lib/api-zod/src/generated/types/noteInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Ambition Classes coaching institute website
 * OpenAPI spec version: 0.1.0
 */
import type { NoteClass } from './noteClass';
import type { NoteSubject } from './noteSubject';

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  class: NoteClass;
  subject: NoteSubject;
  /** @nullable */
  pdfObjectPath?: string | null;
}