# lib/api-zod/src/generated/types/noteSubject.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Ambition Classes coaching institute website
 * OpenAPI spec version: 0.1.0
 */

export type NoteSubject = typeof NoteSubject[keyof typeof NoteSubject];


export const NoteSubject = {
  Mathematics: 'Mathematics',
  Science: 'Science',
  English: 'English',
  Hindi: 'Hindi',
  SST: 'SST',
  AI: 'AI',
  Sanskrit: 'Sanskrit',
} as const;