# lib/api-zod/src/generated/types/uploadUrlResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Ambition Classes coaching institute website
 * OpenAPI spec version: 0.1.0
 */
import type { UploadUrlRequest } from './uploadUrlRequest';

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}